Support for H3 server, negotiated by ALPN
