+++ picoquic/h3zero.c
@@ -730,394 +730,3 @@ int h3zero_send_initial_settings(picoquic_cnx_t * cnx, uint64_t stream_id) {
         sizeof(h3zero_default_setting_frame), 0);
     return ret;
 }
-
-
-/*
- * Create and delete server side connection context
- */
-static h3zero_server_callback_ctx_t * h3zero_server_callback_create_context()
-{
-    h3zero_server_callback_ctx_t* ctx = (h3zero_server_callback_ctx_t*)
-        malloc(sizeof(h3zero_server_callback_ctx_t));
-
-    if (ctx != NULL) {
-        memset(ctx, 0, sizeof(h3zero_server_callback_ctx_t));
-        ctx->first_stream = NULL;
-        ctx->buffer = (uint8_t*)malloc(H3ZERO_RESPONSE_MAX);
-        if (ctx->buffer == NULL) {
-            free(ctx);
-            ctx = NULL;
-        }
-        else {
-            ctx->buffer_max = H3ZERO_RESPONSE_MAX;
-        }
-    }
-
-    return ctx;
-}
-
-static void h3zero_server_callback_delete_context(h3zero_server_callback_ctx_t* ctx)
-{
-    h3zero_server_stream_ctx_t * stream_ctx;
-
-    while ((stream_ctx = ctx->first_stream) != NULL) {
-        ctx->first_stream = stream_ctx->next_stream;
-        free(stream_ctx);
-    }
-
-    if (ctx->buffer != NULL) {
-        free(ctx->buffer);
-        ctx->buffer = NULL;
-    }
-
-    free(ctx);
-}
-
-static h3zero_server_stream_ctx_t * h3zero_find_or_create_stream(
-    picoquic_cnx_t* cnx,
-    uint64_t stream_id,
-    h3zero_server_callback_ctx_t* ctx,
-    int should_create) 
-{
-    h3zero_server_stream_ctx_t * stream_ctx = ctx->first_stream;
-
-    /* if stream is already present, check its state. New bytes? */
-    while (stream_ctx != NULL && stream_ctx->stream_id != stream_id) {
-        stream_ctx = stream_ctx->next_stream;
-    }
-
-    if (stream_ctx == NULL && should_create) {
-        stream_ctx = (h3zero_server_stream_ctx_t*)
-            malloc(sizeof(h3zero_server_stream_ctx_t));
-        if (stream_ctx == NULL) {
-            /* Could not handle this stream */
-            picoquic_reset_stream(cnx, stream_id, H3ZERO_INTERNAL_ERROR);
-        }
-        else {
-            memset(stream_ctx, 0, sizeof(h3zero_server_stream_ctx_t));
-            stream_ctx->next_stream = ctx->first_stream;
-            ctx->first_stream = stream_ctx;
-            stream_ctx->stream_id = stream_id;
-        }
-    }
-
-    return stream_ctx;
-}
-
-/*
- * Incoming data call back.
- * Create context if not yet present.
- * Create stream context if not yet present.
- * Different behavior for unidir and bidir.
- */
-
-static char const * h3zero_default_page = "\
-<!DOCTYPE HTML PUBLIC \"-//IETF//DTD HTML 2.0//EN\">\r\n<HTML>\r\n<HEAD>\r\n<TITLE>\
-Picoquic HTTP 3 service\
-</TITLE>\r\n</HEAD><BODY>\r\n\
-<h1>Simple HTTP 3 Responder</h1>\r\n\
-<p>GET / returns this text</p>\r\n\
-<p>Get /NNNNN returns txt document of length NNNNN bytes(decimal)</p>\r\n\
-<p>Any other command will result in an error, and an empty response.</p>\r\n\
-<h1>Enjoy!</h1>\r\n\
-</BODY></HTML>\r\n";
-
-static int h3zero_server_parse_path(const uint8_t * path, size_t path_length, uint32_t * echo_size)
-{
-    int ret = 0;
-
-    *echo_size = 0;
-    if (path == NULL || path_length == 0 || path[0] != '/') {
-        ret = -1;
-    }
-    else if (path_length > 1) {
-        uint32_t x = 0;
-        for (size_t i = 1; i < path_length; i++) {
-            if (path[i] < '0' || path[i] > '9') {
-                ret = -1;
-                break;
-            }
-            x *= 10;
-            x += path[i] - '0';
-        }
-
-        if (ret == 0) {
-            *echo_size = x;
-        }
-    }
-
-    return ret;
-}
-
-static int h3zero_server_parse_request_frame(
-    picoquic_cnx_t* cnx, 
-    h3zero_server_stream_ctx_t * stream_ctx)
-{
-    int ret = 0;
-
-    uint64_t frame_length;
-    uint8_t * bytes = stream_ctx->frame;
-    uint8_t * bytes_max = bytes + stream_ctx->received_length;
-    uint8_t buffer[1024]; /* used to compose the response */
-
-    /* Parse frame length, verify length */
-    size_t ll = picoquic_varint_decode(stream_ctx->frame, stream_ctx->received_length, &frame_length);
-
-    /* Verify frame type = request header */
-    if (ll == 0) {
-        ret = picoquic_reset_stream(cnx, stream_ctx->stream_id, H3ZERO_INCOMPLETE_REQUEST);
-    }
-    else if (ll + 1 + frame_length < stream_ctx->received_length){
-        ret = picoquic_reset_stream(cnx, stream_ctx->stream_id, H3ZERO_INCOMPLETE_REQUEST);
-    }
-    else if (stream_ctx->frame[ll] != h3zero_frame_header) {
-        ret = picoquic_reset_stream(cnx, stream_ctx->stream_id, H3ZERO_UNEXPECTED_FRAME);
-    }
-    else {
-        /* Parse request header, verify length */
-        h3zero_header_parts_t parts;
-
-        bytes = h3zero_parse_qpack_header_frame(bytes + ll + 1, bytes_max, &parts);
-
-        if (bytes == NULL || bytes != bytes_max) {
-            ret = picoquic_reset_stream(cnx, stream_ctx->stream_id, 
-                H3ZERO_MALFORMED_FRAME(h3zero_frame_header));
-        }
-        else {
-            uint8_t * o_bytes = &buffer[0]; 
-            uint8_t * o_bytes_max = o_bytes + sizeof(buffer);
-            size_t response_length = 0;
-
-            o_bytes += 2; /* reserve two bytes for frame length */
-            *o_bytes++ = h3zero_frame_header;
-
-            /* Parse path */
-            if (parts.method != h3zero_method_get) {
-                /* No such method supported -- error 405, header include "allow GET" */
-                o_bytes = h3zero_create_bad_method_header_frame(o_bytes, o_bytes_max);
-            }
-            else if (h3zero_server_parse_path(parts.path, parts.path_length, &stream_ctx->echo_length) != 0) {
-                /* If unknown, 404 */
-                o_bytes = h3zero_create_not_found_header_frame(o_bytes, o_bytes_max);
-            }
-            else {
-                /* If known, create response header frame */
-                o_bytes = h3zero_create_response_header_frame(o_bytes, o_bytes_max,
-                    (stream_ctx->echo_length == 0) ? h3zero_content_type_text_html :
-                    h3zero_content_type_text_plain);
-
-                response_length = (stream_ctx->echo_length == 0) ?
-                    strlen(h3zero_default_page) : stream_ctx->echo_length;
-            }
-
-            if (o_bytes != NULL) {
-                size_t header_length = o_bytes - &buffer[3];
-                buffer[0] = (uint8_t)((header_length >> 8) | 0x40);
-                buffer[1] = (uint8_t)(header_length &0xFF);
-
-                if (response_length > 0) {
-                    size_t ld = picoquic_varint_encode(o_bytes, o_bytes_max - o_bytes, response_length);
-
-                    if (ld == 0) {
-                        o_bytes = NULL;
-                    }
-                    else {
-                        o_bytes += ld;
-                        if (o_bytes < o_bytes_max) {
-                            *o_bytes++ = h3zero_frame_data;
-
-                            if (stream_ctx->echo_length == 0) {
-                                size_t test_length = strlen(h3zero_default_page);
-
-                                if (o_bytes + test_length <= o_bytes_max) {
-                                    memcpy(o_bytes, h3zero_default_page, test_length);
-                                    o_bytes += test_length;
-                                }
-                                else {
-                                    o_bytes = 0;
-                                }
-                            }
-                        }
-                        else {
-                            o_bytes = NULL;
-                        }
-                    }
-                }
-
-                if (o_bytes != NULL) {
-                    ret = picoquic_add_to_stream(cnx, stream_ctx->stream_id, 
-                        buffer, o_bytes - buffer,
-                        (stream_ctx->echo_length == 0) ? 1 : 0);
-                    if (ret != 0) {
-                        o_bytes = NULL;
-                    }
-                } 
-                
-                if (o_bytes == NULL){
-                    ret = picoquic_reset_stream(cnx, stream_ctx->stream_id, H3ZERO_INTERNAL_ERROR);
-                }
-                else if (stream_ctx->echo_length != 0) {
-                    ret = picoquic_mark_active_stream(cnx, stream_ctx->stream_id, 1);
-                }
-            }
-        }
-    }
-
-    return ret;
-}
-
-static int h3zero_server_callback_data(
-    picoquic_cnx_t* cnx,
-    uint64_t stream_id, uint8_t* bytes, size_t length,
-    picoquic_call_back_event_t fin_or_event,
-    h3zero_server_callback_ctx_t* ctx)
-{
-    int ret = 0;
-    h3zero_server_stream_ctx_t * stream_ctx = NULL;
-
-    /* Find whether this is bidir or unidir stream */
-    if (IS_BIDIR_STREAM_ID(stream_id)) {
-        /* If client bidir stream, absorb data until end, then
-         * parse the header */
-        if (!IS_CLIENT_STREAM_ID(stream_id)) {
-            /* Should never happen */
-            ret = picoquic_stop_sending(cnx, stream_id, H3ZERO_WRONG_STREAM);
-            picoquic_reset_stream(cnx, stream_id, H3ZERO_WRONG_STREAM);
-        }
-        else {
-            /* Find or create stream context */
-            stream_ctx = h3zero_find_or_create_stream(cnx, stream_id, ctx, 1);
-
-            if (stream_ctx == NULL) {
-                ret = picoquic_stop_sending(cnx, stream_id, H3ZERO_INTERNAL_ERROR);
-
-                if (ret == 0) {
-                    ret = picoquic_reset_stream(cnx, stream_id, H3ZERO_INTERNAL_ERROR);
-                }
-            }
-            else {
-                /* Push received bytes at selected offset */
-                if (stream_ctx->received_length + length > sizeof(stream_ctx->frame)) {
-                    /* Too long, unexpected */
-                }
-                else {
-                    memcpy(&stream_ctx->frame[stream_ctx->received_length], bytes, length);
-                    stream_ctx->received_length += length;
-
-                    if (fin_or_event == picoquic_callback_stream_fin) {
-                        /* Parse the request header and process it. */
-                        ret = h3zero_server_parse_request_frame(cnx, stream_ctx);
-                    }
-                }
-            }
-        }
-    }
-    else {
-        /* TODO: If unidir stream, check what type of stream */
-        /* TODO: If this is a control stream, and setting is not received yet,
-         * wait for the setting frame, then process it and move the
-         * state to absorbing.*/
-         /* TODO: Beside control stream, we also absorb the push streams and
-          * the reserved streams (*1F). In fact, we implement that by
-          * just switching the state to absorbing. */
-          /* For now, do nothing, just ignore the data */
-    }
-
-    return ret;
-}
-
-int h3zero_server_callback_prepare_to_send(picoquic_cnx_t* cnx,
-    uint64_t stream_id, void* context, size_t space,
-    h3zero_server_callback_ctx_t* ctx)
-{
-
-    int ret = -1;
-    h3zero_server_stream_ctx_t * stream_ctx = h3zero_find_or_create_stream(cnx, stream_id, ctx, 0);
-
-    if (stream_ctx == NULL) {
-        ret = picoquic_reset_stream(cnx, stream_id, H3ZERO_INTERNAL_ERROR);
-    } else if (stream_ctx->echo_sent < stream_ctx->echo_length){
-        uint8_t * buffer;
-        size_t available = stream_ctx->echo_length - stream_ctx->echo_sent;
-        int is_fin = 1;
-
-        if (available > space) {
-            available = space;
-            is_fin = 0;
-        }
-
-        buffer = picoquic_provide_stream_data_buffer(context, available, is_fin, !is_fin);
-        if (buffer != NULL) {
-            /* TODO: fill buffer with some text */
-            memset(buffer, 0x5A, available);
-            stream_ctx->echo_sent += (uint32_t) available;
-            ret = 0;
-        }
-    }
-
-    return ret;
-}
-
-/*
- * Server call back.
- */
-int h3zero_server_callback(picoquic_cnx_t* cnx,
-    uint64_t stream_id, uint8_t* bytes, size_t length,
-    picoquic_call_back_event_t fin_or_event, void* callback_ctx)
-{
-    int ret = 0;
-    h3zero_server_callback_ctx_t* ctx = (h3zero_server_callback_ctx_t*)callback_ctx;
-    h3zero_server_stream_ctx_t* stream_ctx = NULL;
-
-    if (ctx == NULL) {
-        ctx = h3zero_server_callback_create_context();
-        if (ctx == NULL) {
-            /* cannot handle the connection */
-            picoquic_close(cnx, PICOQUIC_ERROR_MEMORY);
-            return -1;
-        }
-        else {
-            picoquic_set_callback(cnx, h3zero_server_callback, ctx);
-        }
-    }
-
-    switch (fin_or_event) {
-    case picoquic_callback_stream_data:
-    case picoquic_callback_stream_fin:
-        /* Data arrival on stream #x, maybe with fin mark */
-        ret = h3zero_server_callback_data(cnx, stream_id, bytes, length, fin_or_event, ctx);
-        break;
-    case picoquic_callback_stream_reset: /* Client reset stream #x */
-    case picoquic_callback_stop_sending: /* Client asks server to reset stream #x */
-        /* TODO: special case for uni streams. */
-        stream_ctx = h3zero_find_or_create_stream(cnx, stream_id, ctx, 0);
-        if (stream_ctx != NULL) {
-            stream_ctx->status = h3zero_server_stream_status_finished;
-        }
-        picoquic_reset_stream(cnx, stream_id, 0);
-        break;
-    case picoquic_callback_stateless_reset:
-    case picoquic_callback_close: /* Received connection close */
-    case picoquic_callback_application_close: /* Received application close */
-        h3zero_server_callback_delete_context(ctx);
-        picoquic_set_callback(cnx, h3zero_server_callback, NULL);
-        break;
-    case picoquic_callback_stream_gap:
-        /* Gap indication, when unreliable streams are supported */
-        stream_ctx = h3zero_find_or_create_stream(cnx, stream_id, ctx, 0);
-        if (stream_ctx != NULL) {
-            stream_ctx->status = h3zero_server_stream_status_finished;
-        }
-        picoquic_stop_sending(cnx, stream_id, H3ZERO_INTERNAL_ERROR);
-        break;
-    case picoquic_callback_prepare_to_send:
-        /* Used for active streams */
-        ret = h3zero_server_callback_prepare_to_send(cnx, stream_id, (void*)bytes, length, ctx);
-        break;
-    default:
-        /* unexpected */
-        break;
-    }
-
-    return ret;
-}
