+++ picoquic/picoquic.vcxproj
@@ -137,6 +137,7 @@
   <ItemGroup>
     <ClCompile Include="cubic.c" />
     <ClCompile Include="democlient.c" />
+    <ClCompile Include="demoserver.c" />
     <ClCompile Include="fnv1a.c" />
     <ClCompile Include="frames.c" />
     <ClCompile Include="h3zero.c" />
@@ -159,6 +160,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="democlient.h" />
+    <ClInclude Include="demoserver.h" />
     <ClInclude Include="fnv1a.h" />
     <ClInclude Include="h3zero.h" />
     <ClInclude Include="picohash.h" />
