+++ picoquic_t/picoquic_t.c
@@ -155,7 +155,9 @@ static const picoquic_test_def_t test_table[] = {
     { "h3zero_integer", h3zero_integer_test },
     { "h3zero_parse_qpack", h3zero_parse_qpack_test },
     { "h3zero_prepare_qpack", h3zero_prepare_qpack_test },
-    { "h3zero_server", h3zero_server_test }
+    { "h3zero_server", h3zero_server_test },
+    { "h09_server", h09_server_test },
+    { "generic_server", generic_server_test}
 };
 
 static size_t const nb_tests = sizeof(test_table) / sizeof(picoquic_test_def_t);
