+++ picoquictest/picoquictest.h
@@ -154,6 +154,8 @@ int h3zero_integer_test();
 int h3zero_parse_qpack_test();
 int h3zero_prepare_qpack_test();
 int h3zero_server_test();
+int h09_server_test();
+int generic_server_test();
 
 #ifdef __cplusplus
 }
